+++ CHANGELOG.md
@@ -5,6 +5,15 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [2.1.1] - 2024-04-10
+
+### Changed
+
+- Changed order of CloudFormation parameters to emphasize the Security Control playbook
+- Changed default for all playbooks other than SC to 'no'
+- Updated descriptions of playbook parameters
+- Updated architecture diagram
+
 ## [2.1.0] - 2024-03-28
 
 ### Added